requirements.txt
discord
requests
nest-asyncio
python-dotenv
praw